+++ tests/libcxx/host/host.cpp
@@ -41,7 +41,7 @@ void Test(oe_enclave_t* enclave)
 
 OE_OCALL void ocall_exit(uint64_t arg)
 {
-    exit(arg);
+    exit(static_cast<int>(arg));
 }
 
 void* EnclaveThread(void* args)
